+++ sdk/src/main/java/ly/count/android/sdk/messaging/CountlyPush.java
@@ -555,6 +555,11 @@ public class CountlyPush {
                                 msg.recordAction(activity, 0);
                                 dialog.dismiss();
 
+                                if (countlyConfigPush.notificationButtonURLHandler != null && countlyConfigPush.notificationButtonURLHandler.onClick(msg.link().toString())) {
+                                    Countly.sharedInstance().L.d("[CountlyPush, displayDialog] Link handled by custom URL handler, skipping default link opening.");
+                                    return;
+                                }
+
                                 try {
                                     Intent i = new Intent(Intent.ACTION_VIEW, msg.link());
                                     i.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
@@ -596,18 +601,25 @@ public class CountlyPush {
             DialogInterface.OnClickListener listener = new DialogInterface.OnClickListener() {
                 @Override
                 public void onClick(DialogInterface dialog, int which) {
+                    dialog.dismiss();
+                    
+                    boolean isPositiveButtonPressed = (which == DialogInterface.BUTTON_POSITIVE);
+                    if (countlyConfigPush.notificationButtonURLHandler != null && countlyConfigPush.notificationButtonURLHandler.onClick(msg.buttons().get(isPositiveButtonPressed ? 1 : 0).link().toString())) {
+                        Countly.sharedInstance().L.d("[CountlyPush, dialog button onClick] Link handled by custom URL handler, skipping default link opening.");
+                        return;
+                    }
+
                     try {
-                        msg.recordAction(context, which == DialogInterface.BUTTON_POSITIVE ? 2 : 1);
-                        Intent intent = new Intent(Intent.ACTION_VIEW, msg.buttons().get(which == DialogInterface.BUTTON_POSITIVE ? 1 : 0).link());
+                        msg.recordAction(context, isPositiveButtonPressed ? 2 : 1);
+                        Intent intent = new Intent(Intent.ACTION_VIEW, msg.buttons().get(isPositiveButtonPressed ? 1 : 0).link());
                         Bundle bundle = new Bundle();
                         bundle.putParcelable(EXTRA_MESSAGE, msg);
                         intent.putExtra(EXTRA_MESSAGE, bundle);
-                        intent.putExtra(EXTRA_ACTION_INDEX, which == DialogInterface.BUTTON_POSITIVE ? 2 : 1);
+                        intent.putExtra(EXTRA_ACTION_INDEX, isPositiveButtonPressed ? 2 : 1);
                         context.startActivity(intent);
                     } catch (Exception ex) {
-                        Countly.sharedInstance().L.e("[CountlyPush, dialog button onClick] Encountered issue while clicking on button #[" + which + "] [" + ex.toString() + "]");
+                        Countly.sharedInstance().L.e("[CountlyPush, dialog button onClick] Encountered issue while clicking on button #[" + which + "] [" + ex + "]");
                     }
-                    dialog.dismiss();
                 }
             };
             builder.setNeutralButton(msg.buttons().get(0).title(), listener);